增加预言机相关功能

+++ contracts/interface/IPriceFeed.sol
@@ -9,5 +9,5 @@ interface IPriceFeed {
     function aggregator() external view returns (address);
     function latestAnswer() external view returns (int256);
     function latestRound() external view returns (uint80);
-    function getRoundData(uint80 roundId) external view returns (uint80, int256, uint256, uint256, uint80);
+    function getRoundData(uint256 roundId) external view returns (uint256, int256, uint256, uint256, uint80);
 }
